requirements.txt
flet[fastapi]
requests
uvicorn
python-dotenv